Add WavefrontSingleMetric function + Code refactoring
The client only exposes two functions: `Wavefront` which will flush the
entire `go-metrics` registry based on the `FlushInterval` configuration
Duration, and `WavefrontOnce` that will do the same but only once.

This commit refactors the code to introduce a new
`WavefrontSingleMetric` method to easily push a single metric to
Wafefront without registering it to the `go-metrics` registry. This way,
metrics can be submited more sparsely and in a more granular way.